+++ blocks/solana/src/lib.rs
@@ -1,5 +1,6 @@
 mod blocks;
 mod counters;
+mod instructions;
 mod keys;
 mod rewards;
 mod sinks;
